requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.2.1
django-allauth==65.8.1
django-cleanup==9.0.0
django-crispy-forms==2.4
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
pillow==11.2.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
